+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 The changelog for `Superwall`. Also see the [releases](https://github.com/superwall-me/Superwall-Android/releases) on GitHub.
 
+## 1.0.0-alpha.14
+
+### Enhancements
+
+- Adds `trigger_session_id` to Superwall Events.
+- Resets the scroll position of the paywall on close.
+
+### Fixes
+
+- Fixes issue where an invalid currency code on a device would crash the app when trying to retrieve products.
+
 ## 1.0.0-alpha.13
 
 ### Fixes